Revert "feat: upgrade template structure"
