--- src/components/component1.ts
@@ -1,3 +1,0 @@
-export function createPage1Content(): string {
-  return `<div>This is the content of page 1</div>`;
-}
