--- src/components/component2.ts
@@ -1,3 +1,0 @@
-export function createPage2Content(): string {
-  return `<div>This is the content of page 2</div>`;
-}
