--- src/components/home.ts
@@ -1,7 +1,0 @@
-export function createHomePage(): string {
-  return `
-    <div>
-    <h1>Ubiquity TypeScript Template Home</h1>
-    </div>
-  `;
-}
