--- src/shared/types.ts
@@ -1,4 +1,0 @@
-export interface GlobalState {
-  isLoading: boolean;
-  data: Record<string, unknown>;
-}
