+++ static/index.html
@@ -7,15 +7,7 @@
     <link rel="stylesheet" href="style.css" />
   </head>
   <body>
-    <nav id="nav-selector">
-      <a href="#">Home</a>
-      <a href="#/page1">Page 1</a>
-      <a href="#/page2">Page 2</a>
-    </nav>
-
-    <div id="content-area">
-      <!-- This is where the content from pages will be rendered -->
-    </div>
-    <script type="module" src="dist/src/main.js"></script>
+    <h1>Ubiquity TypeScript Template</h1>
+    <script type="module" src="dist/main.js"></script>
   </body>
 </html>
