+++ static/main.ts
@@ -1,0 +1,11 @@
+export async function mainModule() {
+  console.log(`Hello from mainModule`);
+}
+
+mainModule()
+  .then(() => {
+    console.log("mainModule loaded");
+  })
+  .catch((error) => {
+    console.error(error);
+  });
