--- static/page1.html
@@ -1,3 +1,0 @@
-<div class="content">
-  <h1>Ubiquity TypeScript Template Page 1</h1>
-</div>
