--- static/page2.html
@@ -1,3 +1,0 @@
-<div class="content">
-  <h1>Ubiquity TypeScript Template Page 2</h1>
-</div>
