+++ tests/main.test.ts
@@ -1,4 +1,4 @@
-import { mainModule } from "../src/main";
+import { mainModule } from "../static/main";
 import { db } from "./__mocks__/db";
 import { server } from "./__mocks__/node";
 import usersGet from "./__mocks__/users-get.json";
